Task X
Added CSS to allow the text to alternate left and right looking like a chat bubble. Made sure there was code to take the text from the user input field, and set a backup in case the field is empty that makes it generate random text. Then I set the correct class that allows the text to alternate from left to right using the CSS I added using jQuery. Finally using jQuery I appended a new div to the output using the alignmentClass added and made sure there was a line that cleared all of the user input field so that it was not reusing the same text over and over.