Lab 13
I experimented in the console.log and eventually on the js file itself. I had some issues with my code that I thought were extreme, so I had saved a copy of my original attempt and attempted to start again, but eventually decided to go back to my original method and debug. The error was a small capitalization fix that I found while looking at the code on a google doc.